wala is now safe and sound!
- scripts to run only wala
- changing WALA-Ave options such that the analysis yields sound results
there was an issue related to loading/storing to arrays that are of zero
length. In particular the array in hsqldb Table.indexList was
initialzied as indexList = new Index[0] and then elements were added to
it through reflective calls to Array.newInstance and System.arraycopy.
WALA had some checks on ZeroLengthArrayInNode that basically ignored
them. This screwed up WALA-Ave completely for hsaldb but it should be
fixed now.